Make UART with idle line interrupt generic for both IBUS and GPS

+++ software/MIMXRT1011_FlightControllerRtos/source/controllers/attitude_controller.cpp
@@ -14,7 +14,7 @@
  */
 static constexpr float degToRad(float deg)
 {
-    return deg * static_cast<float>(std::numbers::pi_v<double>) / 180.0f;
+    return deg * std::numbers::pi_v<float> / 180.0f;
 }
 
 /* --------------------------------------------------------------------------
